+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Platforms/Android/Mediation/Tapjoy/TapjoyClient.cs
@@ -24,7 +24,9 @@ namespace GoogleMobileAds.Android.Mediation.Tapjoy
     public class TapjoyClient : ITapjoyClient
     {
         private static TapjoyClient instance = new TapjoyClient();
-        private TapjoyClient() {}
+        private TapjoyClient() { }
+
+        private const string TAPJOY_CLASS_NAME = "com.tapjoy.Tapjoy";
 
         public static TapjoyClient Instance
         {
@@ -36,19 +38,26 @@ namespace GoogleMobileAds.Android.Mediation.Tapjoy
 
         public void SetUserConsent(string consentString)
         {
-            AndroidJavaClass tapjoy = new AndroidJavaClass ("com.tapjoy.Tapjoy");
-
-            MonoBehaviour.print ("Calling 'Tapjoy.setUserConsent()' with argument: '" + consentString + "'");
-            tapjoy.CallStatic ("setUserConsent", consentString);
+            AndroidJavaClass tapjoy = new AndroidJavaClass(TAPJOY_CLASS_NAME);
+            AndroidJavaObject tapjoyPrivacyPolicy =
+                    tapjoy.CallStatic<AndroidJavaObject>("getPrivacyPolicy");
+            tapjoyPrivacyPolicy.Call("setUserConsent", consentString);
         }
 
         public void SubjectToGDPR(bool gdprApplicability)
         {
-            AndroidJavaClass tapjoy = new AndroidJavaClass ("com.tapjoy.Tapjoy");
+            AndroidJavaClass tapjoy = new AndroidJavaClass(TAPJOY_CLASS_NAME);
+            AndroidJavaObject tapjoyPrivacyPolicy =
+                    tapjoy.CallStatic<AndroidJavaObject>("getPrivacyPolicy");
+            tapjoyPrivacyPolicy.Call("setSubjectToGDPR", gdprApplicability);
+        }
 
-            string parameterString = (gdprApplicability == true ? "true" : "false");
-            MonoBehaviour.print ("Calling 'Tapjoy.subjectToGDPR()' with argument: " + parameterString);
-            tapjoy.CallStatic ("subjectToGDPR", gdprApplicability);
+        public void SetUSPrivacy(string privacyString)
+        {
+            AndroidJavaClass tapjoy = new AndroidJavaClass(TAPJOY_CLASS_NAME);
+            AndroidJavaObject tapjoyPrivacyPolicy =
+                    tapjoy.CallStatic<AndroidJavaObject>("getPrivacyPolicy");
+            tapjoyPrivacyPolicy.Call("setUSPrivacy", privacyString);
         }
     }
 }