+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Platforms/Mediation/TapjoyClientFactory.cs
@@ -22,13 +22,13 @@ namespace GoogleMobileAds.Mediation
 {
     public class TapjoyClientFactory
     {
-        public static ITapjoyClient TapjoyInstance()
+        public static ITapjoyClient CreateTapjoyClient()
         {
             #if UNITY_EDITOR
             return new GoogleMobileAds.Common.Mediation.Tapjoy.DummyClient();
             #elif UNITY_ANDROID
             return GoogleMobileAds.Android.Mediation.Tapjoy.TapjoyClient.Instance;
-            #elif (UNITY_5 && UNITY_IOS) || UNITY_IPHONE
+            #elif UNITY_IOS
             return GoogleMobileAds.iOS.Mediation.Tapjoy.TapjoyClient.Instance;
             #else
             return new GoogleMobileAds.Common.Mediation.Tapjoy.DummyClient();