+++ source/plugin/Assets/Plugins/iOS/GADUTypes.h
@@ -82,6 +82,9 @@ typedef const void *GADUTypeRequestConfigurationRef;
 /// Type representing a GADUTypeResponseInfoRef
 typedef const void *GADUTypeResponseInfoRef;
 
+/// Type representing a GADUTypeAdapterResponseInfoRef type
+typedef const void *GADUTypeAdapterResponseInfoRef;
+
 /// Type representing a AdError type
 typedef const void *GADUTypeErrorRef;
 
@@ -149,6 +152,13 @@ typedef void (*GADUAdViewWillLeaveApplicationCallback)(GADUTypeBannerClientRef *
 typedef void (*GADUAdViewPaidEventCallback)(GADUTypeBannerClientRef *bannerClient, int precision,
                                             int64_t value, const char *currencyCode);
 
+/// Callback for when an ad reports an impression.
+typedef void (*GADUAdViewImpressionCallback)(GADUTypeBannerClientRef *bannerClient);
+
+/// Callback for when an ad reports a click.
+typedef void (*GADUAdViewClickCallback)(GADUTypeBannerClientRef *bannerClient);
+
+
 // MARK: - GADUInterstitial
 
 /// Callback for when a interstitial ad request was successfully loaded.
@@ -172,6 +182,10 @@ typedef void (*GADUInterstitialAdDidDismissFullScreenContentCallback)(
 
 /// Callback when an interstitial ad has made an impression.
 typedef void (*GADUInterstitialAdDidRecordImpressionCallback)(
+    GADUTypeInterstitialRef *interstitialClient);
+
+/// Callback when an interstitial ad has made a click.
+typedef void (*GADUInterstitialAdDidRecordClickCallback)(
     GADUTypeInterstitialRef *interstitialClient);
 
 /// Callback when an interstitial ad is estimated to have earned money.