+++ mediation/AppLovin/CHANGELOG.md
@@ -1,6 +1,26 @@
 ## AppLovin Unity Mediation Plugin Changelog
 
-#### Version 6.3.1 (In Progress)
+#### Version 6.4.3 (In Progress)
+- Supports [AppLovin Android adapter version 11.5.5.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/applovin/CHANGELOG.md#version-11550).
+- Supports [AppLovin iOS adapter version 11.5.5.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/AppLovin/CHANGELOG.md#version-11550).
+- Built and tested with the Google Mobile Ads Unity Plugin version 7.3.1.
+
+#### [Version 6.4.2](https://dl.google.com/googleadmobadssdk/mediation/unity/applovin/AppLovinUnityAdapter-6.4.2.zip)
+- Supports [AppLovin Android adapter version 11.5.4.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/applovin/CHANGELOG.md#version-11540).
+- Supports [AppLovin iOS adapter version 11.5.4.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/AppLovin/CHANGELOG.md#version-11540).
+- Built and tested with the Google Mobile Ads Unity Plugin version 7.3.0.
+
+#### [Version 6.4.1](https://dl.google.com/googleadmobadssdk/mediation/unity/applovin/AppLovinUnityAdapter-6.4.1.zip)
+- Supports [AppLovin Android adapter version 11.5.3.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/applovin/CHANGELOG.md#version-11531).
+- Supports [AppLovin iOS adapter version 11.5.3.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/AppLovin/CHANGELOG.md#version-11530).
+- Built and tested with the Google Mobile Ads Unity Plugin version 7.3.0.
+
+#### [Version 6.4.0](https://dl.google.com/googleadmobadssdk/mediation/unity/applovin/AppLovinUnityAdapter-6.4.0.zip)
+- Supports [AppLovin Android adapter version 11.5.2.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/applovin/CHANGELOG.md#version-11520).
+- Supports [AppLovin iOS adapter version 11.5.1.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/AppLovin/CHANGELOG.md#version-11510).
+- Built and tested with the Google Mobile Ads Unity Plugin version 7.2.0.
+
+#### [Version 6.3.1](https://dl.google.com/googleadmobadssdk/mediation/unity/applovin/AppLovinUnityAdapter-6.3.1.zip)
 - Added the `AppLovin.SetDoNotSell(bool)` method to support CCPA requirements.
 - Supports [AppLovin Android adapter version 11.4.4.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/applovin/CHANGELOG.md#version-11440).
 - Supports [AppLovin iOS adapter version 11.4.3.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/AppLovin/CHANGELOG.md#version-11430).